Added Adyen Profile test
