+++ uisdk/src/androidTest/java/com/karhoo/uisdk/booking/AdyenBookingTests.kt
@@ -5,6 +5,9 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
 import androidx.test.rule.GrantPermissionRule
 import com.github.tomakehurst.wiremock.junit.WireMockRule
+import com.karhoo.karhootraveller.common.preferences
+import com.karhoo.karhootraveller.menu.menu
+import com.karhoo.karhootraveller.profile.user.userProfile
 import com.karhoo.sdk.api.model.TripInfo
 import com.karhoo.uisdk.common.Launch
 import com.karhoo.uisdk.common.serverRobot
@@ -13,7 +16,6 @@ import com.karhoo.uisdk.screen.booking.BookingActivity
 import com.karhoo.uisdk.util.TestData.Companion.ADYEN
 import com.karhoo.uisdk.util.TestData.Companion.ADYEN_PROVIDER
 import com.karhoo.uisdk.util.TestData.Companion.ADYEN_PUBLIC_KEY
-import com.karhoo.uisdk.util.TestData.Companion.BRAINTREE_PROVIDER
 import com.karhoo.uisdk.util.TestData.Companion.DESTINATION_TRIP
 import com.karhoo.uisdk.util.TestData.Companion.DRIVER_TRACKING
 import com.karhoo.uisdk.util.TestData.Companion.LONG
@@ -24,8 +26,10 @@ import com.karhoo.uisdk.util.TestData.Companion.REVERSE_GEO_SUCCESS
 import com.karhoo.uisdk.util.TestData.Companion.TRIP
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER_NO_NUMBER_PLATE
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_STATUS_DER
+import com.karhoo.uisdk.util.TestData.Companion.USER_INFO_ADYEN
 import com.karhoo.uisdk.util.TestData.Companion.VEHICLES_ASAP
 import com.karhoo.uisdk.util.TestData.Companion.setUserInfo
+import com.schibsted.spain.barista.rule.flaky.AllowFlaky
 import com.schibsted.spain.barista.rule.flaky.FlakyTestRule
 import org.junit.After
 import org.junit.Before
@@ -95,6 +99,31 @@ class AdyenBookingTests : Launch {
             sleep(LONG)
         } result {
             checkDriverDetails()
+        }
+    }
+
+    /**
+     * Given: I have no card linked to my account
+     * When:  I am on the profile screen
+     * Then:  I can see: First Name, Last Name, email, country code, mobile number, add card
+     * button is not visible
+     **/
+    @Test
+    @AllowFlaky(attempts = 10)
+    fun fullCheckProfilePageAdyenUser() {
+        preferences {
+            setUserPreferenceAdyen(USER_INFO_ADYEN)
+        }
+        booking(this) {
+            pressMenuButton()
+        }
+        menu {
+            clickOnProfileButton()
+        }
+        userProfile {
+            waitFor(MEDIUM)
+        } result {
+            fullScreenCheckCardRegisteredAdyen()
         }
     }
